Rename labels to properties and refactor code
The ".js" file extension isn't neccessary in require statements.
