+++ src/webvowl/js/elements/nodes/RectangularNode.js
@@ -1,5 +1,5 @@
-var BaseNode = require("./BaseNode.js");
-var drawTools = require("./drawTools.js")();
+var BaseNode = require("./BaseNode");
+var drawTools = require("./drawTools")();
 
 module.exports = (function () {
 
@@ -78,7 +78,7 @@ module.exports = (function () {
 			}
 			drawTools.appendRectangularClass(parentElement, that.width(), that.height(), cssClasses, that.labelForCurrentLanguage());
 
-			textBlock = require("../../util/textElement.js")(parentElement);
+			textBlock = require("../../util/textElement")(parentElement);
 			textBlock.addText(that.labelForCurrentLanguage());
 
 			that.addMouseListeners();
