+++ src/webvowl/js/elements/nodes/implementations/OwlNothing.js
@@ -1,4 +1,4 @@
-var OwlThing = require("./OwlThing.js");
+var OwlThing = require("./OwlThing");
 
 module.exports = (function () {
