+++ src/webvowl/js/elements/nodes/nodeMap.js
@@ -1,17 +1,17 @@
 var nodes = [];
-nodes.push(require("./implementations/ExternalClass.js"));
-nodes.push(require("./implementations/OwlClass.js"));
-nodes.push(require("./implementations/OwlComplementOf.js"));
-nodes.push(require("./implementations/OwlDeprecatedClass.js"));
-nodes.push(require("./implementations/OwlEquivalentClass.js"));
-nodes.push(require("./implementations/OwlIntersectionOf.js"));
-nodes.push(require("./implementations/OwlNothing.js"));
-nodes.push(require("./implementations/OwlThing.js"));
-nodes.push(require("./implementations/OwlUnionOf.js"));
-nodes.push(require("./implementations/RdfsClass.js"));
-nodes.push(require("./implementations/RdfsDatatype.js"));
-nodes.push(require("./implementations/RdfsLiteral.js"));
-nodes.push(require("./implementations/RdfsResource.js"));
+nodes.push(require("./implementations/ExternalClass"));
+nodes.push(require("./implementations/OwlClass"));
+nodes.push(require("./implementations/OwlComplementOf"));
+nodes.push(require("./implementations/OwlDeprecatedClass"));
+nodes.push(require("./implementations/OwlEquivalentClass"));
+nodes.push(require("./implementations/OwlIntersectionOf"));
+nodes.push(require("./implementations/OwlNothing"));
+nodes.push(require("./implementations/OwlThing"));
+nodes.push(require("./implementations/OwlUnionOf"));
+nodes.push(require("./implementations/RdfsClass"));
+nodes.push(require("./implementations/RdfsDatatype"));
+nodes.push(require("./implementations/RdfsLiteral"));
+nodes.push(require("./implementations/RdfsResource"));
 
 var map = d3.map(nodes, function (Prototype) {
 	return new Prototype().type();
