+++ src/webvowl/js/elements/properties/implementations/OwlDisjointWith.js
@@ -1,9 +1,9 @@
-var BaseLabel = require("../BaseLabel.js");
+var BaseProperty = require("../BaseProperty");
 
 module.exports = (function () {
 
 	var o = function (graph) {
-		BaseLabel.apply(this, arguments);
+		BaseProperty.apply(this, arguments);
 
 		var label = "Disjoint With";
 		// Disallow overwriting the label
@@ -17,7 +17,7 @@ module.exports = (function () {
 			.styleClass("disjointwith")
 			.type("owl:disjointWith");
 	};
-	o.prototype = Object.create(BaseLabel.prototype);
+	o.prototype = Object.create(BaseProperty.prototype);
 	o.prototype.constructor = o;
 
 	return o;
