+++ src/webvowl/js/elements/properties/implementations/RdfsSubClassOf.js
@@ -1,9 +1,9 @@
-var BaseLabel = require("../BaseLabel.js");
+var BaseProperty = require("../BaseProperty");
 
 module.exports = (function () {
 
 	var o = function (graph) {
-		BaseLabel.apply(this, arguments);
+		BaseProperty.apply(this, arguments);
 
 		var that = this,
 			superDrawFunction = that.draw,
@@ -25,7 +25,7 @@ module.exports = (function () {
 			.styleClass("subclass")
 			.type("rdfs:subClassOf");
 	};
-	o.prototype = Object.create(BaseLabel.prototype);
+	o.prototype = Object.create(BaseProperty.prototype);
 	o.prototype.constructor = o;
 
 	return o;
