+++ src/webvowl/js/elements/properties/implementations/SetOperatorProperty.js
@@ -1,9 +1,9 @@
-var BaseLabel = require("../BaseLabel.js");
+var BaseProperty = require("../BaseProperty");
 
 module.exports = (function () {
 
 	var o = function (graph) {
-		BaseLabel.apply(this, arguments);
+		BaseProperty.apply(this, arguments);
 
 		this.markerType("special")
 			.labelVisible(false)
@@ -11,7 +11,7 @@ module.exports = (function () {
 			.styleClass("setoperatorproperty")
 			.type("setOperatorProperty");
 	};
-	o.prototype = Object.create(BaseLabel.prototype);
+	o.prototype = Object.create(BaseProperty.prototype);
 	o.prototype.constructor = o;
 
 	return o;
