+++ src/webvowl/js/util/languageTools.js
@@ -1,4 +1,4 @@
-var constants = require("./constants.js")();
+var constants = require("./constants")();
 
 /**
  * Encapsulates methods which return a label in a specific language for a preferred language.
